\name{cells20}
\alias{cells20}
\docType{data}
\title{
Indices of mesh cells for spherical numerical integration
}
\description{
Indices of mesh cells for spherical numerical integration.
}
\format{
  A data frame with 8000 rows where each one corresponds to the indices (rows) in the data set \code{points20} that characterizes a cell.
  \describe{
    \item{\code{V1}}{Index of the first cell vertex.}
    \item{\code{V2}}{Index of the second cell vertex.}
    \item{\code{V3}}{Index of the third cell vertex.}
  }
}
\keyword{internal}
