% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_apply_pipeline}
\alias{hdir_apply_pipeline}
\title{Apply Image Algorithms to a Set of Images}
\usage{
hdir_apply_pipeline(paths, ...)
}
\arguments{
\item{paths}{a vector of paths to the images}

\item{...}{inputs passed to the function}
}
\value{
nothing
}
\description{
Applies a set of Python libraries to the set of images.
}
