% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_cache_get}
\alias{hdir_cache_get}
\title{Download URL}
\usage{
hdir_cache_get(url, cache_dir, page = NULL, force = FALSE)
}
\arguments{
\item{url}{string of the url to grab}

\item{cache_dir}{string of the cache directory}

\item{page}{name of the page; used only for progress message}

\item{force}{if set, will ignore and overwrite current page in the cache}
}
\value{
A response object from the httr package.
}
\description{
Downloads the content of a URL, with a cache implemented to avoiding
re-downloading the same data multiple times.
}
