% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_network_metrics}
\alias{hdir_network_metrics}
\title{Add network metadata and layout information}
\usage{
hdir_network_metrics(edges, node_name = "id", directed = FALSE)
}
\arguments{
\item{edges}{data frame describing the edges}

\item{node_name}{name of the node id column; defaults to "id"}

\item{directed}{logical; is the graph directed?}
}
\value{
a list with two elements, each containing a tibble.
}
\description{
Given a data frame with two columns giving edges, runs a suite
of network algorithms over the resulting graph.
}
