% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_pixels_long}
\alias{hdir_pixels_long}
\title{Create Long Format of Image Pixels}
\usage{
hdir_pixels_long(paths, names = NULL)
}
\arguments{
\item{paths}{a vector of paths to the images}

\item{names}{a vector of names for the images that
serve as a key to the images; will be
equal to the paths if not provided}
}
\value{
a tibble object with one row for every pixel
                 in every image.
}
\description{
Give a set of paths to images, this function will create
a long-format version of the pixels in all the images, 
with own row per pixel. The rows contain the red, green,
and blue pixel intensities as well as the derivted hue,
saturation and value, and a hex description of the color.
}
