% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_plot_images}
\alias{hdir_plot_images}
\title{Save Set of Images as a Grid}
\usage{
hdir_plot_images(paths, ncol = NULL, nrow = NULL)
}
\arguments{
\item{paths}{a vector of paths to the images}

\item{ncol}{number of columns in the output}

\item{nrow}{number of rows in the output}
}
\value{
an array of pixel intensities.
}
\description{
Given a set of paths to images, this function will create
an array of images that can be saved as a JPEG or PNG file.
}
