% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_text_tfidf}
\alias{hdir_text_tfidf}
\title{Construct the TF-IDF Matrix from Annotation or Data Frame}
\usage{
hdir_text_tfidf(
  object,
  min_df = 0.1,
  max_df = 0.9,
  max_features = 10000,
  doc_var = "doc_id",
  token_var = "lemma",
  vocabulary = NULL
)
}
\arguments{
\item{object}{a data frame containing an identifier for the document
(set with \code{doc_var}) and token (set with
\code{token_var})}

\item{min_df}{the minimum proportion of documents a token
should be in to be included in the vocabulary}

\item{max_df}{the maximum proportion of documents a token
should be in to be included in the vocabulary}

\item{max_features}{the maximum number of tokens in the vocabulary}

\item{doc_var}{character vector. The name of the column in
\code{object} that contains the document ids. Defaults
to "doc_id".}

\item{token_var}{character vector. The name of the column in
\code{object} that contains the tokens. Defaults to
"lemma".}

\item{vocabulary}{character vector. The vocabulary set to use in
constructing the matrices. Will be computed
within the function if set to \code{NULL}. When
supplied, the options \code{min_df}, \code{max_df},
and \code{max_features} are ignored.}
}
\value{
a tibble in wide format with term frequencies and tf-idf values.
}
\description{
Given annotations, this function returns the term-frequency inverse
document frequency (tf-idf) matrix from the extracted lemmas.
}
