% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_tidy_angle_distance}
\alias{hdir_tidy_angle_distance}
\title{Create a data frame of pairwise angle distances}
\usage{
hdir_tidy_angle_distance(x, item_name = "document")
}
\arguments{
\item{x}{the matrix to compute similarities between rows.}

\item{item_name}{name to give to the set of rows; defaults to "document"}
}
\value{
a tibble in long format, with one row for each pair of observations.
}
\description{
Given a matrix with observations in the rows labeled with the row names
and numeric features in the columns, this produces a data frame of the
angle distances between all pairs of observations.
}
\examples{
mat <- matrix(rnorm(26 * 5), ncol = 5)
rownames(mat) <- letters
hdir_tidy_angle_distance(mat)

}
