% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_tidy_pca}
\alias{hdir_tidy_pca}
\title{Create a data frame of principal components}
\usage{
hdir_tidy_pca(x, n = 3, scale = TRUE, item_name = "document")
}
\arguments{
\item{x}{the matrix to compute principal components from.}

\item{n}{number of components to compute}

\item{scale}{logical; should columns by scaled before PCA}

\item{item_name}{name to give to the set of rows; defaults to "document"}
}
\value{
a tibble with one row for each observation.
}
\description{
Given a matrix with observations in the rows labeled with the row names
and numeric features in the columns, this produces a data frame of the
principal components for each observation.
}
\examples{
mat <- matrix(rnorm(26 * 5), ncol = 5)
rownames(mat) <- letters
hdir_tidy_pca(mat)

}
