% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_tidy_umap}
\alias{hdir_tidy_umap}
\title{Create a data frame of umap components}
\usage{
hdir_tidy_umap(x, n = 2, random_state = 1, item_name = "document", ...)
}
\arguments{
\item{x}{the matrix to compute principal components from.}

\item{n}{number of components to compute}

\item{random_state}{integer to set the random state of the algorithm}

\item{item_name}{name to give to the set of rows; defaults to "document"}

\item{...}{other options passed to the umap function}
}
\value{
a tibble with one row for each observation.
}
\description{
Given a matrix with observations in the rows labeled with the row names
and numeric features in the columns, this produces a data frame of the
UMAP projection coordinates for each observation.
}
\examples{
mat <- matrix(rnorm(26 * 5), ncol = 5)
rownames(mat) <- letters
hdir_tidy_umap(mat)

}
