\name{points10}
\alias{points10}
\docType{data}
\title{
Coordinates of mesh vertices for spherical numerical integration
}
\description{
Coordinates of mesh vertices for spherical numerical integration.
}
\format{
  A data frame with 1002 rows where each one corresponds to the coordinates of vertices that characterize a cell.
  \describe{
    \item{\code{V1}}{Coordinate of the first cell vertex.}
    \item{\code{V2}}{Coordinate of the second cell vertex.}
    \item{\code{V3}}{Coordinate of the third cell vertex.}
  }
}
\keyword{internal}
