% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{theme_tufte}
\alias{theme_tufte}
\title{Minimal theme perfect for presentations}
\usage{
theme_tufte()
}
\value{
a ggplot2 theme object.
}
\description{
A minimalistic theme based on Tufte's *Visual Display of Quantitative Information*.
}
