\name{trapezoidal.rule}
\alias{trapezoidal.rule}
 \title{
Trapezoidal rule for numerical integration over a threshold}
\description{
This functions implements the trapezoidal rule for numerical integration of a function \eqn{f} over a threshold.}
\usage{
trapezoidal.rule(f, step, y)
}
 \arguments{
  \item{f}{
Function to be integrated.}
  \item{step}{
Numeric value of the distance between points for integration.}
  \item{y}{
Numeric value of the threshold.}
}

\value{
A numeric value of the integral.
}

\author{
Paula Saavedra-Nieves and Rosa M. Crujeiras.
}

\keyword{internal}
