% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{hdir_apply_pipeline}
\alias{hdir_apply_pipeline}
\title{Apply Image Algorithms to a Set of Images}
\usage{
hdir_apply_pipeline(
  paths,
  fnames = paths,
  outdir = tempdir(),
  verbose = TRUE,
  prefix = "",
  embed_type = NULL
)
}
\arguments{
\item{paths}{a vector of paths to the images}

\item{fnames}{an optional set of filenames to use the metadata;
will be set to the full path names if not provided}

\item{outdir}{directory to store the output in; otherwise will 
store in a temporary directory}

\item{verbose}{should the function report its progress; defaults
to TRUE}

\item{prefix}{optional prefix for the output filenames}

\item{embed_type}{if running the embed annotation, what type of values
should be returned. Either "dimred", "full", or "all".
Defaults to "dimred" for more than 50 inputs and "full"
otherwise.}
}
\value{
directory where the output was saved
}
\description{
Applies a set of Python libraries to the set of images.
}
